npyz

Module npz

Utilities for working with npz files.

NPZ files are a format used by numpy to store archives of multiple arrays. The format itself is just a .zip file containing files with a .npy extension.

When the "npz" feature is enabled, this module provides adapters around the [zip][::zip] crate that read and write NPZ files.

Even without the "npz" feature, this module always provides a set of utility functions for converting between array names and filenames inside the zip file. This allows you to use your own zip library. The examples directory includes an example of how to read and write an npz without the feature.

Functions

array_name_from_file_name

Get the name of the array that would correspond to the given filename inside a zip file.

file_name_from_array_name

Get the filename in a zip that np.savez would use for a keyword argument.

file_name_from_index

Get the filename in a zip that would be used for an array supplied as a positional argument to np.savez.