Trait npyz::Deserialize

pub trait Deserialize: Sized {
    type TypeReader: TypeRead<Value = Self>;

    // Required method
    fn reader(dtype: &DType) -> Result<Self::TypeReader, DTypeError>;
}

Trait that permits reading a type from an .npy file.

Complete documentation of all types that implement this trait is available on the type_matchup_docs module.

Examples of types that implement this:

Primitive integers, floats, Complex (with the "complex" feature)

Owned containers (Vec<u8>, String)

This trait is derivable when enabling the "derive" feature. This makes it easier to work with structured arrays.

For an example of how to implement this manually, see Vector5 in the roundtrip test.

Required Associated Types

type TypeReader: TypeRead<Value = Self>

Think of this as like a for<R: io::Read> Fn(R) -> io::Result<Self>.

There is no closure-like sugar for these; you must manually define a type that implements TypeRead.

Required Methods

fn reader(dtype: &DType) -> Result<Self::TypeReader, DTypeError>

Get a function that deserializes a single data field at a time.

The purpose of the dtype arugment is to allow e.g. specifying a length for string types, or the endianness for integers.

Errors

Returns Err if the DType is not compatible with Self.

Object Safety

This trait is not object safe.