Trait npyz::Serialize

pub trait Serialize {
    type TypeWriter: TypeWrite<Value = Self>;

    // Required method
    fn writer(dtype: &DType) -> Result<Self::TypeWriter, DTypeError>;
}

Trait that permits writing a type to an .npy file.

Complete documentation of all types that implement this trait is available on the type_matchup_docs module.

Examples of types that implement this:

Primitive integers, floats, Complex (with the "complex" feature)

Slice types ([u8], str)

This trait is derivable when enabling the "derive" feature. This makes it easier to work with structured arrays.

For an example of how to implement this manually, see Vector5 in the roundtrip test.

Required Associated Types

type TypeWriter: TypeWrite<Value = Self>

Think of this as some sort of for<W: io::Write> Fn(W, &Self) -> io::Result<()>.

There is no closure-like sugar for these; you must manually define a type that implements TypeWrite.

Required Methods

fn writer(dtype: &DType) -> Result<Self::TypeWriter, DTypeError>

Get a function that serializes a single data field at a time.

The purpose of the dtype arugment is to allow e.g. specifying a length for string types, or the endianness for integers. The derivable AutoSerialize trait is able to supply many types with a reasonable default.

Errors

Returns Err if the DType is not compatible with Self.

Object Safety

This trait is not object safe.