npyz

Trait TypeRead

pub trait TypeRead {
    type Value;

    // Required method
    fn read_one<R: Read>(&self, bytes: R) -> Result<Self::Value>
       where Self: Sized;
}

Like some sort of for<R: io::Read> Fn(R) -> io::Result<T>.

To obtain one of these, use the Deserialize trait.

For an example of how to implement this manually, see Vector5 in the roundtrip test.

Trait objects

dyn TypeRead has no object-safe methods. If you need dynamic polymorphism, use dyn TypeReadDyn instead.

Required Associated Types

type Value

Type returned by the function.

Required Methods

fn read_one<R: Read>(&self, bytes: R) -> Result<Self::Value>
where Self: Sized,

The function.