Type Alias npyz::OutFile

pub type OutFile<Row> = NpyWriter<Row, BufWriter<File>>;

👎Deprecated since 0.5.0: Doesn’t carry its weight. Use to_file_1d instead, or replicate the original behavior with Builder::new().default_dtype().begin_1d(std::io::BufWriter::new(std::fs::File::create(path)?))

NpyWriter that writes an entire file.

Aliased Type

struct OutFile<Row> { /* private fields */ }

Implementations

impl<Row: AutoSerialize> OutFile<Row>

pub fn open<P: AsRef<Path>>(path: P) -> Result<Self>

👎Deprecated since 0.5.0: Doesn’t carry its weight. Use to_file_1d instead, or replicate the original behavior with Builder::new().default_dtype().begin_1d(std::io::BufWriter::new(std::fs::File::create(path)?))

Create a file, using the default format for the given type.

impl<Row: Serialize> OutFile<Row>

pub fn close(self) -> Result<()>

👎Deprecated since 0.5.0: use .finish() instead

Finish writing the file and close it. Alias for NpyWriter::finish.

If omitted, the file will be closed on drop automatically, ignoring any errors encountered during the process.