npyz::write_options

Trait HasShape

pub trait HasShape { }

Indicates that a Writer options type includes a shape.

If you are receiving the following compiler error:

the trait bound `npyz::WriteOptions<i64>: npyz::write_options::HasShape` is not satisfied

Then it is most likely because you are missing a call to WriterBuilder::shape.