npyz::write_options

Trait HasWriter

pub trait HasWriter {
    type Writer;
}

Indicates that a Writer options type includes an output stream.

If you are receiving the following compiler error:

the trait bound `npyz::WriteOptions<i64>: npyz::write_options::HasWriter` is not satisfied

Then it is most likely because you are missing a call to WriterBuilder::writer.

Required Associated Types

type Writer

The type of the output stream.