Function npyz::npz::array_name_from_file_name

pub fn array_name_from_file_name(path_in_zip: &str) -> Option<&str>

Get the name of the array that would correspond to the given filename inside a zip file.

This tries to match the behavior of numpy’s own npz-loading behavior:

Case sensitive (only .npy files, not .NPY files)

Allows weird characters like / and .. Makes no attempt to normalize paths.

Treats an interior null as the end of the path. Notice that this means that multiple different filenames could produce the same array name in a maliciously constructed zip.

Returns None if numpy would not consider the file to be an array.

This function does not require any cargo features.