Struct npyz::TypeStr

pub struct TypeStr { /* private fields */ }

Represents an Array Interface type-string.

This is more or less the DType of a scalar type. Exposes a FromStr impl for construction, and a Display impl for writing.

use npyz::TypeStr;

let ts = "|i1".parse::<TypeStr>()?;

assert_eq!(format!("{}", ts), "|i1");
assert_eq!(ts.endianness(), npyz::Endianness::Irrelevant);
assert_eq!(ts.type_char(), npyz::TypeChar::Int);
assert_eq!(ts.size_field(), 1);

Implementations

impl TypeStr

pub fn endianness(&self) -> Endianness

Extract the endianness character from the type string.

pub fn type_char(&self) -> TypeChar

Extract the type character from the type string.

For most (but not all!) types, this is the number of bytes that a single value occupies. For the U type, it is the number of code units.

pub fn size_field(&self) -> u64

Extract the “size” field from the type string. This is the number that appears after the type character.

For most (but not all!) types, this is the number of bytes that a single value occupies. For the U type, it is the number of code units.

pub fn time_units(&self) -> Option<TimeUnits>

Extract the time units, if this type string has any. Only TypeChar::TimeDelta and TypeChar::DateTime have time units.

pub fn num_bytes(&self) -> Option<usize>

Get the number of bytes for a single value.

If this value would overflow the platform’s usize type, returns None.