Module npyz::write_options

Types and traits related to the implementation of WriteOptions.

There are numerous types and traits involved in the implementation of WriteOptions. Most of them are not that important; their purpose is to implement a form of typestate. E.g. you are forbidden from calling WriterBuilder::begin_nd until you have called WriterBuilder::shape, WriterBuilder::dtype, and WriterBuilder::writer.

The most important things from this module (WriteOptions and WriterBuilder) are exported from the root of the crate, so you typically do not need to look in here.

Structs

WithDType
Return type of WriterBuilder::dtype. It represents a config with a known dtype.

WithShape
Return type of WriterBuilder::shape. It represents a config with a known array shape.

WithWriter
Return type of WriterBuilder::writer. It represents a config with a known output stream.

WriteOptions
Represents an almost-empty configuration for an NpyWriter.

Traits

HasDType
Indicates that a Writer options type includes a DType.

HasShape
Indicates that a Writer options type includes a shape.

HasWriter
Indicates that a Writer options type includes a writer.

WriterBuilder
Trait that provides methods on WriteOptions.