npyz::write_options

Trait MissingWriter

pub trait MissingWriter { }

Indicates that a Writer options type does not yet have an output stream.

If you are receiving the following compiler error:

the trait bound `npyz::write_options::HasWriter<...>: npyz::write_options::MissingWriter` is not satisfied

Then it is most likely because you are calling WriterBuilder::writer on a builder that already has a writer. For instance, [NpzWriter::array][crate::npz::NpzWriter::array] comes with a writer, so you do not need to add one.